Signature checking / Fix warnings / sync win11 input handling

+++ CChromaEditorLibrary/CChromaEditorLibrary.cpp
@@ -572,11 +572,11 @@ void CMainViewDlg::RefreshGrid()
 			vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 			int currentFrame = _mEdit1D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame1D& frame = frames[currentFrame];
 				int id = 0;
@@ -603,11 +603,11 @@ void CMainViewDlg::RefreshGrid()
 			vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 			int currentFrame = _mEdit2D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame2D& frame = frames[currentFrame];
 				int id = 0;
@@ -669,11 +669,11 @@ void CMainViewDlg::RefreshFrames()
 			vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 			int currentFrame = _mEdit1D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame1D& frame = frames[currentFrame];
 				sprintf_s(bufferDuration, "%f", frame.Duration);
@@ -686,11 +686,11 @@ void CMainViewDlg::RefreshFrames()
 			vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 			int currentFrame = _mEdit2D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame2D& frame = frames[currentFrame];
 				sprintf_s(bufferDuration, "%f", frame.Duration);
@@ -1422,11 +1422,11 @@ void CMainViewDlg::OnBnClickedButtonColor(UINT nID)
 						vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 						int currentFrame = _mEdit1D.GetCurrentFrame();
 						if (currentFrame < 0 ||
-							currentFrame >= frames.size())
+							currentFrame >= (int)frames.size())
 						{
 							currentFrame = 0;
 						}
-						if (currentFrame < frames.size())
+						if (currentFrame < (int)frames.size())
 						{
 							FChromaSDKColorFrame1D& frame = frames[currentFrame];
 							int i = index;
@@ -1475,11 +1475,11 @@ void CMainViewDlg::OnBnClickedButtonColor(UINT nID)
 						vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 						int currentFrame = _mEdit2D.GetCurrentFrame();
 						if (currentFrame < 0 ||
-							currentFrame >= frames.size())
+							currentFrame >= (int)frames.size())
 						{
 							currentFrame = 0;
 						}
-						if (currentFrame < frames.size())
+						if (currentFrame < (int)frames.size())
 						{
 							FChromaSDKColorFrame2D& frame = frames[currentFrame];
 							int i = index / maxColumn;
@@ -1730,11 +1730,11 @@ void CMainViewDlg::OnBnClickedButtonClear()
 			vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 			int currentFrame = _mEdit1D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame1D& frame = frames[currentFrame];
 				frame.Colors = ChromaSDKPlugin::GetInstance()->CreateColors1D(device);
@@ -1748,11 +1748,11 @@ void CMainViewDlg::OnBnClickedButtonClear()
 			vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 			int currentFrame = _mEdit2D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame2D& frame = frames[currentFrame];
 				frame.Colors = ChromaSDKPlugin::GetInstance()->CreateColors2D(device);
@@ -1780,11 +1780,11 @@ void CMainViewDlg::OnBnClickedButtonFill()
 			vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 			int currentFrame = _mEdit1D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame1D& frame = frames[currentFrame];
 				for (int i = 0; i < maxLeds; ++i)
@@ -1803,11 +1803,11 @@ void CMainViewDlg::OnBnClickedButtonFill()
 			vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 			int currentFrame = _mEdit2D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame2D& frame = frames[currentFrame];
 				for (int i = 0; i < maxRow; ++i)
@@ -1841,11 +1841,11 @@ void CMainViewDlg::OnBnClickedButtonFillBlank()
 		vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 		int currentFrame = _mEdit1D.GetCurrentFrame();
 		if (currentFrame < 0 ||
-			currentFrame >= frames.size())
+			currentFrame >= (int)frames.size())
 		{
 			currentFrame = 0;
 		}
-		if (currentFrame < frames.size())
+		if (currentFrame < (int)frames.size())
 		{
 			FChromaSDKColorFrame1D& frame = frames[currentFrame];
 			for (int i = 0; i < maxLeds; ++i)
@@ -1867,11 +1867,11 @@ void CMainViewDlg::OnBnClickedButtonFillBlank()
 		vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 		int currentFrame = _mEdit2D.GetCurrentFrame();
 		if (currentFrame < 0 ||
-			currentFrame >= frames.size())
+			currentFrame >= (int)frames.size())
 		{
 			currentFrame = 0;
 		}
-		if (currentFrame < frames.size())
+		if (currentFrame < (int)frames.size())
 		{
 			FChromaSDKColorFrame2D& frame = frames[currentFrame];
 			for (int i = 0; i < maxRow; ++i)
@@ -1908,11 +1908,11 @@ void CMainViewDlg::OnBnClickedButtonDarken()
 		vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 		int currentFrame = _mEdit1D.GetCurrentFrame();
 		if (currentFrame < 0 ||
-			currentFrame >= frames.size())
+			currentFrame >= (int)frames.size())
 		{
 			currentFrame = 0;
 		}
-		if (currentFrame < frames.size())
+		if (currentFrame < (int)frames.size())
 		{
 			FChromaSDKColorFrame1D& frame = frames[currentFrame];
 			for (int i = 0; i < maxLeds; ++i)
@@ -1938,11 +1938,11 @@ void CMainViewDlg::OnBnClickedButtonDarken()
 		vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 		int currentFrame = _mEdit2D.GetCurrentFrame();
 		if (currentFrame < 0 ||
-			currentFrame >= frames.size())
+			currentFrame >= (int)frames.size())
 		{
 			currentFrame = 0;
 		}
-		if (currentFrame < frames.size())
+		if (currentFrame < (int)frames.size())
 		{
 			FChromaSDKColorFrame2D& frame = frames[currentFrame];
 			for (int i = 0; i < maxRow; ++i)
@@ -1983,11 +1983,11 @@ void CMainViewDlg::OnBnClickedButtonLighten()
 		vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 		int currentFrame = _mEdit1D.GetCurrentFrame();
 		if (currentFrame < 0 ||
-			currentFrame >= frames.size())
+			currentFrame >= (int)frames.size())
 		{
 			currentFrame = 0;
 		}
-		if (currentFrame < frames.size())
+		if (currentFrame < (int)frames.size())
 		{
 			FChromaSDKColorFrame1D& frame = frames[currentFrame];
 			for (int i = 0; i < maxLeds; ++i)
@@ -2013,11 +2013,11 @@ void CMainViewDlg::OnBnClickedButtonLighten()
 		vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 		int currentFrame = _mEdit2D.GetCurrentFrame();
 		if (currentFrame < 0 ||
-			currentFrame >= frames.size())
+			currentFrame >= (int)frames.size())
 		{
 			currentFrame = 0;
 		}
-		if (currentFrame < frames.size())
+		if (currentFrame < (int)frames.size())
 		{
 			FChromaSDKColorFrame2D& frame = frames[currentFrame];
 			for (int i = 0; i < maxRow; ++i)
@@ -2057,11 +2057,11 @@ void CMainViewDlg::OnBnClickedButtonRandom()
 			vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 			int currentFrame = _mEdit1D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame1D& frame = frames[currentFrame];
 				frame.Colors = ChromaSDKPlugin::GetInstance()->CreateRandomColors1D(device);
@@ -2075,11 +2075,11 @@ void CMainViewDlg::OnBnClickedButtonRandom()
 			vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 			int currentFrame = _mEdit2D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame2D& frame = frames[currentFrame];
 				frame.Colors = ChromaSDKPlugin::GetInstance()->CreateRandomColors2D(device);
@@ -2104,11 +2104,11 @@ void CMainViewDlg::OnBnClickedButtonCopy()
 			vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 			int currentFrame = _mEdit1D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame1D& frame = frames[currentFrame];
 				_mEdit1D.SetCopy(frame);
@@ -2121,11 +2121,11 @@ void CMainViewDlg::OnBnClickedButtonCopy()
 			vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 			int currentFrame = _mEdit2D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame2D& frame = frames[currentFrame];
 				_mEdit2D.SetCopy(frame);
@@ -2148,11 +2148,11 @@ void CMainViewDlg::OnBnClickedButtonPaste()
 			vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 			int currentFrame = _mEdit1D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				frames[currentFrame] = _mEdit1D.GetCopy();
 				RefreshGrid();
@@ -2166,11 +2166,11 @@ void CMainViewDlg::OnBnClickedButtonPaste()
 			vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 			int currentFrame = _mEdit2D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				frames[currentFrame] = _mEdit2D.GetCopy();
 				RefreshGrid();
@@ -2197,11 +2197,11 @@ void CMainViewDlg::OnBnClickedButtonPreview()
 			vector<FChromaSDKColorFrame1D>& frames = _mEdit1D.GetFrames();
 			int currentFrame = _mEdit1D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame1D& frame = frames[currentFrame];
 				FChromaSDKEffectResult result = ChromaSDKPlugin::GetInstance()->CreateEffectCustom1D(device, frame.Colors);
@@ -2219,11 +2219,11 @@ void CMainViewDlg::OnBnClickedButtonPreview()
 			vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 			int currentFrame = _mEdit2D.GetCurrentFrame();
 			if (currentFrame < 0 ||
-				currentFrame >= frames.size())
+				currentFrame >= (int)frames.size())
 			{
 				currentFrame = 0;
 			}
-			if (currentFrame < frames.size())
+			if (currentFrame < (int)frames.size())
 			{
 				FChromaSDKColorFrame2D& frame = frames[currentFrame];
 				FChromaSDKEffectResult result = ChromaSDKPlugin::GetInstance()->CreateEffectCustom2D(device, frame.Colors, frame.Keys);
@@ -2319,11 +2319,11 @@ void CMainViewDlg::OnBnClickedButtonSetKey()
 		vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 		int currentFrame = _mEdit2D.GetCurrentFrame();
 		if (currentFrame < 0 ||
-			currentFrame >= frames.size())
+			currentFrame >= (int)frames.size())
 		{
 			currentFrame = 0;
 		}
-		if (currentFrame < frames.size())
+		if (currentFrame < (int)frames.size())
 		{
 			int id = GetControlSetKeyCombo()->GetCurSel();
 			EChromaSDKKeyboardKey key = (EChromaSDKKeyboardKey)id;
@@ -2348,11 +2348,11 @@ void CMainViewDlg::OnBnClickedButtonSetLed()
 		vector<FChromaSDKColorFrame2D>& frames = _mEdit2D.GetFrames();
 		int currentFrame = _mEdit2D.GetCurrentFrame();
 		if (currentFrame < 0 ||
-			currentFrame >= frames.size())
+			currentFrame >= (int)frames.size())
 		{
 			currentFrame = 0;
 		}
-		if (currentFrame < frames.size())
+		if (currentFrame < (int)frames.size())
 		{
 			int id = GetControlSetLEDCombo()->GetCurSel();
 			EChromaSDKMouseLED led = (EChromaSDKMouseLED)id;
